Implement cobra argument parsing and move logic from main() to cmd.Execute()
